(fix): moved ColorHelper functions to frontend
why were they in backend????
